requirements.txt
# Flask Backend Requirements for Mental Health Companion

# Core Flask
Flask==2.3.3
Flask-CORS==4.0.0
Flask-Limiter==3.5.0

# HTTP Requests
requests==2.31.0

# AI/ML Libraries
google-generativeai==0.3.2

# Database
sqlite3

# Utilities
python-dotenv==1.0.0
gunicorn==21.2.0

# Development
pytest==7.4.3
pytest-flask==1.3.0